realm-model + node + persist: graduate realm-model into the running node — node-hosted admission gate + durable realms across restart on the keystone receipt-chain (driven canary)
The world-substrate deep-read's #1 load-bearing disconnection: realm-model (realms/instances/canonical-identity/committed-catalog/per-realm
membrane/hybrid-PQ identity+guardian recovery) was REAL as a model but CONNECTEDNESS=NONE — own empty [workspace], ZERO non-test reverse deps,
not a workspace member; the running node re-implemented the same concepts AD-HOC in JS + raw agent cells. Its load-bearing dependency (a durable
node-served receipt chain) JUST LANDED (the keystone), so it can graduate.
WIRED (the 3 MUD-SUBSTRATE deliverables): (1) realm-model is a ROOT WORKSPACE MEMBER + default-member (deleted its standalone [workspace]/[patch]
— inherits root); compiles + its 16 tests pass. (2) ADMISSION through realm-model's gate: node now DEPENDS on realm-model; new node/src/realm_
service.rs (NodeRealms) HOSTS a real realm_model::RealmWorld; NodeState holds it (realms field + accessor) + restores it in both constructors;
every realm/instance turn routes through RealmWorld::admit/settle_instance — the catalog check, identity resolution, and scope membrane are realm-
model's, NOT ad-hoc JS. (3) DURABLE across restart on the keystone chain: new persist REALM_LOG table + append_realm_log_entry/load_realm_log/
realm_log_len, built on the EXACT dense/gap/idempotence discipline of the keystone RECEIPT_CHAIN; only admitted ops logged; NodeRealms::restore
replays the dense log through a fresh RealmWorld on boot (mirroring install_receipt_chain_durability); persist stays byte-oriented (ZERO realm-
model dep).
⚑ THE CANARY (node/src/state.rs::realm_created_through_node_survives_restart_and_refuses_false_catalog, driven green): builds a realm THROUGH THE
NODE (mint identity -> bind surface -> create realm -> list ruleset -> open instance -> play 40 -> settle), records the realm receipt head/hoard,
DROPs the NodeState, REOPENS the same data dir, and asserts the realm SURVIVED — same receipt head, hoard=40/epoch=1, surface still resolves to
the canonical id, catalog still committed. AND the DEPLOYED REFUSAL: a turn citing an unlisted ruleset_root returns Refused::RulesetNotInCatalog
+ persists NOTHING (durable REALM_LOG length unchanged). The keystone receipt_chain_and_mmr_head_survive_node_restart + both corrupt-boot fail-
closed tests still pass (no regression). BEFORE: realm_model zero non-test reverse deps. AFTER: node uses it + a realm survives a real restart.
HONEST SCOPE — WIRED vs NAMED RESIDUAL (not laundered): GENUINELY WIRED = node-hosted durable RealmWorld; admission routes through realm-model's
gate (not JS); catalog/identity/membrane are the gate's; durable-across-restart via REALM_LOG replay. NAMED RESIDUAL = the gate is RealmWorld::
admit CALLED BY THE NODE, NOT YET inside the kernel executor's proof_verify.rs, and ruleset_root is NOT YET a field of dregg_turn::Turn (MUD-
SUBSTRATE items 1-2 — additive signed-turn-perimeter changes deliberately not forced from thin context). So a realm turn is a node-hosted,
durable, gate-routed admission, not yet the signed/fee-estimated/proof-carrying kernel Turn the HTTP /turns/submit ingress drives; no public
/realm HTTP ingress yet (reachable via NodeState::realms()); succession/guardian-recovery ops aren't in the durable RealmOp log yet (the canary
covers realm/instance/identity-mint/catalog persistence, not key rotation). (Also fixed realm-model's rot vs dregg_turn::Effect — SetField.index
u64 at HEAD vs the usize it was authored for; a usize<->u64 bridge in world.rs + two tests. ⚠ HEAD's dregg-pq is separately broken [guard_no_
verified_core called-undefined, a wip commit] — flagged to ember, NOT repaired in this commit.)

+++ realm-model/tests/driven.rs
@@ -218,7 +218,7 @@ fn catalog_is_committed_law() {
         ruleset_root: root,
         effects: vec![Effect::SetField {
             cell: inst.id,
-            index: realm_model::instance::field::RESULT,
+            index: realm_model::instance::field::RESULT as u64,
             value: pack_u64(1),
         }],
     };
@@ -287,7 +287,7 @@ fn instance_scope_membrane_holds() {
         ruleset_root: law,
         effects: vec![Effect::SetField {
             cell: realm.id,
-            index: realm_model::realm::field::HOARD,
+            index: realm_model::realm::field::HOARD as u64,
             value: pack_u64(1_000_000),
         }],
     });
